src/main/java/top/kealine/judgehost/testcase/LRUCache.java
package top.kealine.judgehost.testcase;

import java.util.LinkedHashMap;
import java.util.Map;

public class LRUCache<K, V> extends LinkedHashMap<K, V> {
    private final int MAX_CACHE_SIZE;
    private Map.Entry<K, V> removed;

    public LRUCache(int cacheSize) {
        super((int) Math.ceil(cacheSize / 0.75) + 1, 0.75f, true);
        MAX_CACHE_SIZE = cacheSize;
    }

    @Override
    protected boolean removeEldestEntry(Map.Entry eldest) {
        if (size() > MAX_CACHE_SIZE) {
            this.removed = eldest;
            return true;
        } {
            return false;
        }
    }

    public Map.Entry<K, V> getRemoved() {
        Map.Entry<K, V> removed = this.removed;
        this.removed = null;
        return removed;
    }

    @Override
    public String toString() {
        StringBuilder sb = new StringBuilder();
        for (Map.Entry<K, V> entry : entrySet()) {
            sb.append(String.format("%s:%s ", entry.getKey(), entry.getValue()));
        }
        return sb.toString();
    }
}